+++ client/src/components/Header.jsx
@@ -27,13 +27,13 @@ export default function Header() {
                         href="/rent"
                         className="text-base font-medium text-gray-500 hover:text-gray-900"
                     >
-                        Rent
+                        Arrivals
                     </a>
                     <a
-                        href="/lend"
+                        href="/departures"
                         className="text-base font-medium text-gray-500 hover:text-gray-900"
                     >
-                        Lend
+                        Departures
                     </a>
                 </Popover.Group>
                 <div className="hidden md:flex items-center justify-end md:flex-1 lg:w-0">
