+++ client/src/pages/index.js
@@ -1,4 +1,4 @@
+export { default as Departures } from './Departures'
 export { default as Landing } from './Landing'
-export { default as Lend } from './Lend'
-export { default as LendCheckout } from './LendCheckout'
 export { default as Rent } from './Rent'
+export { default as RentCheckout } from './RentCheckout'
